% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cm}
\alias{cm}
\title{Cardiomyocite data.}
\format{
A RDS object.
}
\usage{
cm
}
\description{
This dataset was collected by Elborany et al., 2021 and used in our paper.
}
\examples{
data(cm)
}
\references{

}
\keyword{datasets}
