% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsplit.R
\name{fit_glms}
\alias{fit_glms}
\title{Anna note to self- should I make this called "testDE" or something and include t-tests and Wilcoxon tests as options??
IDK how much I should plan to do for the user.}
\usage{
fit_glms(
  counts,
  latentvar,
  family = "poisson",
  test = "wald",
  offsets = rep(1, NROW(counts))
)
}
\arguments{
\item{counts}{A cell-by-gene matrix of integer counts.}

\item{latentvar}{The latent variable that you wish to test for differential expression across. Must have length equal to number of cells}

\item{family}{The family for the GLM. Typically will be "poisson", "quasipoisson", or "nb". Other options may cause errors.}

\item{test}{Specify the type of test you would like. For now the only choice is "wald".}

\item{offsets}{Specify any offsets that you would like in your model. Typically a vector of logged size factors. Must have length equal to the number of cells.}
}
\description{
Anna note to self- should I make this called "testDE" or something and include t-tests and Wilcoxon tests as options??
IDK how much I should plan to do for the user.
}
