% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogNorm}
\alias{LogNorm}
\title{modified Log normalization}
\usage{
LogNorm(data, scale_factor, eps_train)
}
\arguments{
\item{data}{A sparse gene-by-cell matrix of integer counts}

\item{scale_factor}{a scale factor (usually 10000)}

\item{eps_train}{A double that determines the proportion of information that is allocated to each fold. Default is 1 - 1/folds.}
}
\description{
This is log normalization function based on Seurats normalization method and slightly altered for this package.
Reference: https://github.com/satijalab/seurat/blob/763259d05991d40721dee99c9919ec6d4491d15e/src/data_manipulation.cpp#L113
}
