% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsplit.R
\name{nb.countsplit}
\alias{nb.countsplit}
\title{Negative binomial count splitting}
\usage{
nb.countsplit(X, epsilon = 0.5, overdisps = NULL, folds = 2)
}
\arguments{
\item{X}{A cell-by-gene matrix of integer counts}

\item{epsilon}{The thinning parameter for count splitting. Must be between 0 and 1.}

\item{overdisps}{A vector of length p, where p is the number of columns in X. This vector stores the
gene specific overdispersion parameters. Note that a value of overdisp=infinity corresponds to the Poisson distribution.}

\item{folds}{The number of independent folds of data to return. Currently, the only option is 2.}
}
\description{
Takes one matrix of counts and returns a list containing two matrices of counts: a training set and a test set.
}
\details{
The training set and the test set are independent under the assumption that the original data follow a Negative binomial distribution.
Be sure to see the countsplit tutorial vignette for details of how to use this correctly with existing
This function is still in testing mode. Tutorials for this function are coming soon.
}
