% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc.counts}
\alias{pbmc.counts}
\title{This was downloaded using the instructions at the start of the Seurat clustering tutorial.
It is included here for convenience.}
\format{
A matrix.
}
\usage{
pbmc.counts
}
\description{
This was downloaded using the instructions at the start of the Seurat clustering tutorial.
It is included here for convenience.
}
\examples{
data(pbmc.counts)
}
\keyword{datasets}
