% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cm}
\alias{cm}
\title{Cardiomyocyte data.}
\format{
A RDS object.
}
\usage{
cm
}
\description{
This is a subset of a dataset that was collected by Elborany et al., 2022 and used in our paper. This
subset contains 10,000 cells and more than 20,000 genes.
}
\examples{
data(cm)
}
\keyword{datasets}
