% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cm}
\alias{cm}
\title{Cardiomyocyte data from Elorbany et al., 2022.}
\format{
A RDS object.
}
\usage{
cm
}
\description{
This is a subset of a dataset that was collected by Elborany et al., 2022. This data was also analyzed in Neufeld et al., 2022.
This subset contains 10,000 cells and more than 20,000 genes. The format is a SingleCellExperiment object.
}
\examples{
data(cm)
}
\references{
Elorbany, Reem, et al. "Single-cell sequencing reveals lineage-specific dynamic genetic regulation of gene expression during human cardiomyocyte differentiation." PLoS genetics 18.1 (2022): e1009666.
}
\keyword{datasets}
