% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsplit.R
\name{countsplit}
\alias{countsplit}
\title{Takes one matrix of counts and returns a list containing two matrices of counts: a training set and a test set.}
\usage{
countsplit(X, epsilon = 0.5)
}
\arguments{
\item{X}{A cell-by-gene matrix of integer counts}

\item{epsilon}{The thinning parameter for count splitting. Must be between 0 and 1.}
}
\description{
Be sure to see the countsplitting tutorial vignette for details of how to use this correctly with existing
single cell RNA-seq pipelines.
}
