% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogNorm}
\alias{LogNorm}
\title{modified Log normalization}
\usage{
LogNorm(mat, scale_factor, eps_train)
}
\arguments{
\item{mat}{A sparse gene-by-cell matrix of integer counts}

\item{scale_factor}{a scale factor (usually 10000)}

\item{eps_train}{A double that determines the proportion of information that is allocated to each fold. Default is 1 - 1/folds.}
}
\value{
A sparse matrix of log-normalized counts multiplied by 1/eps_train.
}
\description{
This function performs log(x+1) normalization after divided by the total counts per cell and multiplied by a scale factor, as done in Seurat.
The only difference is that we here multiply by 1/eps_train.
Also this function has been rewritten in RcppArmadillo.
}
