% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc.counts}
\alias{pbmc.counts}
\title{PBMC data from 10X Genomics.}
\format{
A dgCMatrix.
}
\usage{
pbmc.counts
}
\description{
This dataset is downloaded at the start of the Seurat guided clustering tutorial (https://satijalab.org/seurat/articles/pbmc3k_tutorial.html).
It is a dataset of Peripheral Blood Mononuclear Cells (PBMC) which has been made freely available by 10X Genomics.
The dataset contains raw counts for 2,700 single cells that were sequenced on the Illumina NextSeq 500.
}
\examples{
data(pbmc.counts)
}
\keyword{datasets}
